Moved implementation to utils.cc as windows,unix use the same code.
